Add unit tests for discount calculator

+++ LogicLibrary/Models/Products/Product.cs
@@ -3,9 +3,9 @@
     public abstract class Product
     {
         public string Name { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
 
-        protected Product(string name, float price)
+        protected Product(string name, decimal price)
         {
             Name = name;
             Price = price;
